Various updates to bring code up to par for PR
