+++ libs/oidc/admin/ngx/src/lib/modules/users/edit-users/detail-user/detail-user.component.ts
@@ -14,7 +14,7 @@ import {
 import { User, Role, INewRole } from '@tamu-gisc/oidc/common';
 
 @Component({
-  selector: 'detail-user',
+  selector: 'tamu-gisc-detail-user',
   templateUrl: './detail-user.component.html',
   styleUrls: ['./detail-user.component.scss']
 })
@@ -25,7 +25,6 @@ export class DetailUserComponent implements OnInit, OnDestroy {
   public userForm: FormGroup;
   public $clients: Observable<Array<Partial<IClientMetadataResponse>>>;
   public $roles: Observable<Array<Partial<Role>>>;
-
   private _$destroy: Subject<boolean> = new Subject();
 
   constructor(
@@ -111,12 +110,14 @@ export class DetailUserComponent implements OnInit, OnDestroy {
           ...user,
           account: user.account
         });
+
         this.userForm.valueChanges.pipe(debounceTime(1000), takeUntil(this._$destroy)).subscribe((res) => {
           console.log('User', this.userForm.getRawValue());
 
           const updatedUser: Partial<User> = {
             ...this.userForm.getRawValue()
           };
+
           this.userService.updateUser(updatedUser).subscribe((result) => [console.log('Updated details')]);
         });
       });
