+++ libs/oidc/common/src/lib/oidc-common.ts
@@ -10,6 +10,7 @@ export * from './modules/role/role.module';
 export * from './modules/user/user.module';
 
 // services
+export * from './services/account/account.service';
 export * from './services/client-metadata/client-metadata.service';
 export * from './services/role/role.service';
 export * from './services/user/user.service';
