--- libs/oidc/common/src/lib/services/common/common.service.spec.ts
@@ -1,19 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-
-import { CommonService } from './common.service';
-
-describe('CommonService', () => {
-  let service: CommonService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [CommonService]
-    }).compile();
-
-    service = module.get<CommonService>(CommonService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
