+++ libs/oidc/provider/src/lib/modules/oidc/oidc.module.ts
@@ -1,11 +1,10 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
+import { AccountRepo, AccountService, RoleRepo } from '@tamu-gisc/oidc/common';
+
 import { OidcProviderService } from '../../services/provider/provider.service';
 import { OidcController } from '../../controllers/oidc/oidc.controller';
-import { AccountService } from '../../services/account/account.service';
-
-import { AccountRepo, RoleRepo } from '../../../../../common/src/lib/entities/all.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([AccountRepo, RoleRepo])],
