+++ libs/oidc/provider/src/lib/oidc-provider.ts
@@ -8,6 +8,5 @@ export * from './modules/user-login/user-login.module';
 export * from './modules/access-token/access-token.module';
 
 // services
-export * from './services/account/account.service';
 export * from './services/provider/provider.service';
 export * from './services/user-login/user-login.service';
