+++ libs/oidc/provider/src/lib/services/provider/provider.service.ts
@@ -1,11 +1,12 @@
+import { readFileSync, writeFile } from 'fs';
 import { Injectable } from '@nestjs/common';
 
 import { JWK } from 'node-jose';
 import { Configuration, JWKS, Provider, ResourceServer } from 'oidc-provider';
-import { readFileSync, writeFile } from 'fs';
 
-import { AccountService } from '../account/account.service';
-import { OidcAdapter } from '../../adapters/oidc.adapter.new';
+import { AccountService } from '@tamu-gisc/oidc/common';
+
+import { OidcAdapter } from '../../adapters/oidc.adapter';
 
 @Injectable()
 export class OidcProviderService {
@@ -13,7 +14,7 @@ export class OidcProviderService {
   private enableDevLogs = true;
   public provider: Provider;
   public issuerUrl = 'http://localhost:4001';
-  private pathToJWKS = 'idp_keystore.json';
+  private pathToJWKS = 'idp_keystore.json'; // TODO: Get file name / path from environment
 
   constructor(private readonly accountService: AccountService) {
     this.getJWKSFromFile().then((jwks) => {
@@ -115,7 +116,7 @@ export class OidcProviderService {
         },
         devInteractions: { enabled: this.devInteractions }, // defaults to true
         deviceFlow: { enabled: true }, // defaults to false
-        issAuthResp: { enabled: true }, // defaults to false
+        // issAuthResp: { enabled: true }, // Was a property in v6, not v7. Will keep in case Panva adds back. Defaults to false
         registration: {
           enabled: true,
           initialAccessToken: false
@@ -141,32 +142,14 @@ export class OidcProviderService {
             } as ResourceServer;
           },
           defaultResource: (ctx) => {
-            // console.log('defaultResource', ctx);
-            // return 'http://localhost:4204';
             return ctx.request.header.referer;
           }
         },
-
         revocation: { enabled: true } // defaults to false
       },
       async findAccount(ctx, sub, token) {
-        // @param ctx - koa request context
-        // @param sub {string} - account identifier (subject)
-        // @param token - is a reference to the token used for which a given account is being loaded,
-        //   is undefined in scenarios where claims are returned from authorization endpoint
-
         return {
           accountId: sub,
-          // @param use {string} - can either be "id_token" or "userinfo", depending on
-          //   where the specific claims are intended to be put in
-          // @param scope {string} - the intended scope, while oidc-provider will mask
-          //   claims depending on the scope automatically you might want to skip
-          //   loading some claims from external resources or through db projection etc. based on this
-          //   detail or not return them in ID Tokens but only UserInfo and so on
-          // @param claims {object} - the part of the claims authorization parameter for either
-          //   "id_token" or "userinfo" (depends on the "use" param)
-          // @param rejected {Array[String]} - claim names that were rejected by the end-user, you might
-          //   want to skip loading some claims from external resources or through db projection
           async claims(use, scope, claims, rejected) {
             console.log('findAccount -> claims', use, scope, claims, rejected);
             const account = await accountService.accountRepo.findOne({
@@ -174,7 +157,6 @@ export class OidcProviderService {
                 guid: sub
               }
             });
-            // return { use, scope, claims, sub: sub, ...account };
             return { sub: account.guid, name: account.name };
           }
         };
@@ -192,14 +174,13 @@ export class OidcProviderService {
             // TODO: Maybe we just use the client_id attribute instead of a guid, since the client_id is a name that wouldn't change between registrations
             // console.log(ctx, token, jwt);
             // const account = await accountService.get(jwt.payload.sub);
-            // jwt.payload.role = 'user';
+            // jwt.payload.role = 'user'; // TODO: Add back authorization; this is a way to "fudge" it
             return jwt;
           }
         }
       },
       interactions: {
         url(ctx, interaction) {
-          // eslint-disable-line no-unused-vars
           return `/interaction/${interaction.uid}`;
         }
       },
@@ -213,16 +194,6 @@ export class OidcProviderService {
         );
       }
     };
-
-    // Code block not really needed anymore now that we put it in the getJWKSFromFile() method
-    // const keystore = JWK.createKeyStore();
-
-    // await keystore.generate('RSA', 2048, {
-    //   alg: 'RS256',
-    //   use: 'sig'
-    // });
-
-    // const jwks: JWKS = keystore.toJSON(true) as JWKS;
 
     // Some properties shouldn't be stored in git such as the keys and clients, cookies, etc
     return {
