--- libs/oidc/provider/src/lib/types/oidc-provider.type.ts
@@ -1,5 +1,0 @@
-import { Provider } from 'oidc-provider';
-
-export interface OidcProvider extends Provider {
-  setProviderSession(req, res, {});
-}
