+++ package.json
@@ -51,7 +51,6 @@
     "@nestjs/typeorm": "^8.0.0",
     "@turf/centroid": "^6.0.2",
     "@types/luxon": "^2.0.5",
-    "@types/oidc-provider": "^7.8.1",
     "angular-auth-oidc-client": "^13.1.0",
     "angular-bootstrap-datetimepicker": "^2.2.1",
     "angulartics2": "^12.0.0",
@@ -143,6 +142,7 @@
     "@types/mapbox-gl": "^1.9.1",
     "@types/multer": "^1.4.3",
     "@types/node": "14.14.33",
+    "@types/oidc-provider": "^7.8.1",
     "@types/yargs": "^16.0.1",
     "@typescript-eslint/eslint-plugin": "~5.10.0",
     "@typescript-eslint/parser": "~5.10.0",
